+++ CHANGELOG.md
@@ -7,17 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- `tlru.TLRU`, is the TTL instance of `tlru.LRU`. It can be created with `NewTTL`.
+- `tlru.WithClock` can allow customization of clock for `tlru.TLRU`.
 - `lrucore.TTLCore`, is the TTL instance of `lrucore.Core`. It can be created with `NewTTL`.
 - `lrucore.WithClock` can allow customization of clock for `lrucore.TTLCore`.
-- More Examples for `GoDoc`.
+- `lrucore.Shard` interface representing `lrucore.Core` and `lrucore.TTLCore`
+- More Examples for `GoDoc`. 
 
 ### Changed
+- `tlru.LRU[K comparable, V any]` is now `tlru.LRU[K comparable, V any, C lrucore.Shard]`.
 - `lrucore.Core` no longer supports TTL.
+- `Option` has been renamed to `LRUOption`.
 - `CoreOption` has been renamed to `TTLOption`.
+- `tlru.WithClock` cannot be used with `tlru.New`.
 - `lrucore.WithClock` cannot be used with `lrucore.New`.
 
 ### Removed
-- `lrucore.WithTTL` has been removed.
+- `lrucore.WithTTL` and `tlru.WithTTL` has been removed.
 
 ## [0.5.1]
 Link : [Github Release](https://github.com/justpranavrs/tlru/releases/tag/v0.5.1)